+++ Assets/Scripts/Card/CardEffect.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b179bdf978399f443b3a6767b90172c6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
