+++ Assets/Scripts/Card/CardEffect/BuffEffect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 79ab535e2a626934f86b1d85543fcf85
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
