+++ Assets/Scripts/Card/CardEffect/CardEffect.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public abstract class CardEffect : ScriptableObject
+{
+    public abstract void ApplyEffect();
+}
