+++ Assets/Scripts/Card/CardEffect/DamageEffect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8a058d47de8fc2469ca037c2c286ae3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
